pyproject.toml
[tool.poetry]
name = "letsh-api"
version = "0.1.0"
description = ""
authors = ["donghhan <apensia914@gmail.com>"]
readme = "README.md"
packages = [{include = "letsh_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"
django-cities-light = "^3.9.2"
drf-yasg = "^1.21.5"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
